# Injaz_backend_code/src/swagger.json
{
    "swagger":"2.0",
    "info":{
       "description":"This is the swagger node.js API project",
       "version":"1.0.0",
       "title":"Authentication API ES "
    },
...
    "host":"https://api.logicrent.ae",
    "paths":{
...
"/driver/signUp":{
   "post":{
      "summary":"user signUp",
      "description":"creation of user",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"user",
            "description":"user object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/user"
                 
               
            }
         }
      ],
      "responses":{
         "201":{
            "description":"user created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "409":{
            "description":"user alredy exist",
            "sechema":{
               "$ref":"#/definitions/userResponse"
            }
         },
         "500":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
"/driver/login":{
   "post":{
      "summary":"user login",
      "description":"login user ",
      "produces":[
         "application/json"
      ],
      "consumes":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"body",
            "description":"userLogin object",
            "required":true,
            "schema":{
             "type":"object",
             "$ref":"#/definitions/userLogin"
               
         
         }
     }
      ],
      "responses":{
         "201":{
            "description":"user login is done",
            "schema":{
               "type":"object",
                  "$ref":"#/definitions/userResponse"
               
            }
         },
         "401":{
            "description":"Invalid Email or Password",
            "schema":{
               "$ref":"#/definitions/userResponse"
            }
         },
         "404":{
             "description":"user doesnot exist",
             "schema":{
                "$ref":"#/definitions/userResponse"
             }
          },
          "500":{
             "description":"something went wrong please try again",
             "schema":{
                "$ref":"#/definitions/InvalidResponse"
             }
          }
      }
   }
},
"/driver/createNewlist":{
   "post":{
      "summary":"creation of NewList",
      "description":"creation of NewList",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"NewList",
            "description":"NewList object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/newCar"  
            }
         }
      ],
      "responses":{
         "201":{
            "description":"NewList created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "500":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
"/driver/addContactInfo":{
   "post":{
      "summary":"creation of addContactInfo",
      "description":"creation of addContactInfo",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"addContactInfo",
            "description":"addContactInfo object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/addContactInfo"  
            }
         }
      ],
      "responses":{
         "200":{
            "description":"ContactInfo created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "400":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
"/driver/addtradeLicence":{
   "post":{
      "summary":"creation of addtradeLicence",
      "description":"creation of addtradeLicence",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"addtradeLicence",
            "description":"addtradeLicence object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/addtradeLicence"  
            }
         }
      ],
      "responses":{
         "200":{
            "description":"addtradeLicence created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "400":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
"/driver/addCorporateVedio":{
   "post":{
      "summary":"creation of addCorporateVedio",
      "description":"creation of addCorporateVedio",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"addCorporateVedio",
            "description":"addCorporateVedio object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/addCorporateVedio"  
            }
         }
      ],
      "responses":{
         "200":{
            "description":"addCorporateVedio created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "400":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
"/driver/addCharges":{
   "post":{
      "summary":"creation of addCharges",
      "description":"creation of addCharges",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"addCharges",
            "description":"addCharges object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/addCharges"  
            }
         }
      ],
      "responses":{
         "201":{
            "description":"addCharges created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "500":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
...
    },
    "definitions":{
       "userResponse":{
          "type":"object",
          "properties":{
             "status":{
                "type":"integer"
             },
             "message":{
                "type":"string"
             }
             
          }
       },
...
       "user":{
          "type":"object",
          "properties":{
             "id":{
                "type":"objectId"
             },
             "userName":{
                "type":"string"
             },
             "email":{
                "type":"string"
             },
             "password":{
                "type":"string"
             },
             "phoneNumber":{
                "type":"string"
             },
             "address":{
               "type":"string"
             },
             "locality":{
               "type":"string"
             },
             "area":{
               "type":"string"
             },
             "zipcode":{
               "type":"string"
             },
             "city":{
               "type":"string"
             }

          }
        },
          "userLogin":{
            "type":"object",
            "properties":{
              
               "email":{
                  "type":"string"
               },
               "password":{
                  "type":"string"
               }
            }
       },
...
       "InvalidResponse":{
          "type":"object",
          "properties":{
             "statusCode":{
                "type":"string"
             },
             "message":{
                "type":"string"
             }
          }
       },
         "newCar": {
            "type": "object",
            "properties": {
               "name":
               {
                  "type":"string"
               },
               "brand":
               {
                  "type":"string"
               },
               "model":
               {
                  "type":"string"
               },
               "category":
               {
                  "type":"string"
               },
              "year":
               {
                  "type":"string"
               },
               "image":
               {
                  "type":"string"
               },
               "externalImage":{
                  "type": "array",
                  "items": {
                    "type": "string"
                  }

               },
               "internalImage":{
                  "type": "array",
                  "items": {
                    "type": "string"
                  }

               },
               "location":
               {
                  "type":"string"
               },
               "vehicleType":
               {
                  "type":"string"
               },
               "featuredCar":
               {
                  "type":"string"
               },
               "status": 
               {
                  "type":"string"
               },
               "services":{
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
                ,
               "description":
               {
                  "type":"string"
               },
               "actualPriceDaily":
               {
                  "type":"string"
               },
               "interiorColor":{
                  "type":"string"
               },
               "exteriorColor":{
                  "type":"string"
               },
               "discountedPriceDaily":
               {
                  "type":"string"
               },
               "actualPriceWeekly":
               {
                  "type":"string"
               },
               "discountedPriceWeekly":
               {
                  "type":"string"
               },
               "actualPriceMonthly":
               {
                  "type":"string"
               },
               "discountedPriceMonthly":
               {
                  "type":"string"
               },
               "transmission":
               {
                  "type":"string"
               },
               "cruiseControl":
               {
                  "type":"string"
               },
               "engineCapacity":
               {
                  "type":"string"
               },
               "laggageBootCapacity":
               {
                  "type":"string"
               },
               "securityDeposit":
               {
                  "type":"string"
               },
               "cashType":{
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
                ,
               "carFeatures":{
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
                ,
               "unlimitedMileage":
               {
                  "type":"string"
               },
               "paiInsuranceDaily":
               {
                  "type":"string"
               },
               "paiInsuranceMonthly":
               {
                  "type":"string"
               },
               "freeCancellation":
               {
                  "type":"string"
               },
               "freeDelivery30DaysandAbove":
               {
                  "type":"string"
               },
               "customerSupport":
               {
                  "type":"string"
               },
               "scdwPerMonth":
               {
                  "type":"string"
               },
               "paymentType":{
                  "type":"string"
               },
               "engineSize":
               {
                  "type":"string"
               },
               "bluetooth":
               {
                  "type":"string"
               },
               "aux":
               {
                  "type":"string"
               },
               "seater":
               {
                  "type":"string"
               },
               "navigation":
               {
                  "type":"string"
               },
               "parkingSense":{
                  "type":"string"
               },
               "requirementsForUAEResidents":{
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
                ,
               "requirementsForTourists":{
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
                
            }
          },
...
"addContactInfo":{
   "type":"object",
   "properties":{
   "Email":{
      "type":"string"
   },
   "WhatsupNo":{
      "type":"string"
   },
   "SMSNotificationNo":{
      "type":"string"
   },
   "MainAdress":{
      "type":"string"
   },
   "MainAdressArabic":{
      "type":"string"
   },
   "PhoneNo":{
      "type":"string"
   }              
}

},
"addtradeLicence":{
   "type":"object",
   "properties":{
   "UploadTradLicense":{
      "type":"string"
   },
   "LicenseExpiryDate":{
      "type":"string"
   }          
}

},
"addCorporateVedio":{
   "type":"object",
   "properties":{
   "CORPORATE_VIDEO":{
      "type":"string"
   }          
}

}
...
"addCharges": {
   "type": "object",
   "properties": {
      "DefaultCharges":{
         "type": "object",
         "properties": {
          
           
   
         "Spes":{"type":"string"},
         "ExtaMillageCost":{"type":"string"},
         "ExcessClaim":{"type":"string"},
         "MilleagePerDay":{"type":"string"},
         "MilleagePerMonth":{"type":"string"},
         "MilleagePerWeek":{"type":"string"},
         "SecurityDeposite":{"type":"string"},
        "AcceptedIn":{"type":"string"},
        "MinimumCustomers":{"type":"string"},
        "RequiredDriving":{"type":"string"},
       "SpecialNoteFor":{"type":"string"}
         }
   },
   "ExtraCharges":{
      "type": "object",
      "properties": {
   
     "GPSDevice":{"type":"string"},
     "CDWPerDayCost":{"type":"string"},
     "CDWPerMonthCost":{"type":"string"},
     "Salik":{"type":"string"},
    "AdditionalDriverInsurance":{"type":"string"},
     "BabySeat":{"type":"string"}
      }
   }
   }
 },
...
    }
 }